version 0.3 22.02.2023

+++ internal/model/prices.go
@@ -5,5 +5,5 @@ package model
 type Price struct {
 	Name string
 	SellingPrice,
-	PurchasePrice float32
+	PurchasePrice float64
 }
